Refactor card rarity probabilities to use double values for improved accuracy

+++ src/main/java/polycube/polycard/card/Rarity.java
@@ -10,7 +10,7 @@ import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 /// Represents the rarity of a card, including its rarity level, probability of occurrence,
 /// whether it is enchanted, and a description of its effect.
 public record Rarity(
-        RarityLevel rarityLevel, int probability, boolean isEnchanted, String description,
+        RarityLevel rarityLevel, double probability, boolean isEnchanted, String description,
         Multimap<Holder<Attribute>, AttributeModifier> attributeModifiers
 ) {
     public Component getFormattedDescription() {
